Added new function IterableToFormattedString
This new function accepts iterable containers filled with primitive
types and created a formatted string with the container's contents.
An element separator, a line separator, noof max. elements in a line,
noof max. elements to be listed and
a max elements exceeded string can be passed optionally.
(Inspired by the existing CreateInStmt() function)

Change-Id: I76c41d93199f0fb85429d6d435759d00f1a94f38